+++ utils/check-working-tree/__tests__/check-working-tree.test.js
@@ -1,8 +1,10 @@
 "use strict";
 
 jest.mock("@lerna/describe-ref");
+jest.mock("@lerna/collect-uncommitted");
 
 const describeRef = require("@lerna/describe-ref");
+const collectUncommitted = require("@lerna/collect-uncommitted");
 const checkWorkingTree = require("../lib/check-working-tree");
 
 describe("check-working-tree", () => {
@@ -29,13 +31,25 @@ describe("check-working-tree", () => {
 
   it("rejects when working tree has uncommitted changes", async () => {
     describeRef.mockResolvedValueOnce({ isDirty: true });
+    collectUncommitted.mockResolvedValueOnce(["AD file"]);
 
     try {
       await checkWorkingTree();
     } catch (err) {
       expect(err.message).toMatch("Working tree has uncommitted changes");
+      expect(err.message).toMatch("\nAD file");
     }
 
+    expect.assertions(2);
+  });
+
+  it("passes cwd to collectUncommitted when working tree has uncommitted changes", async () => {
+    describeRef.mockResolvedValueOnce({ isDirty: true });
+    try {
+      await checkWorkingTree({ cwd: "foo" });
+    } catch (err) {
+      expect(collectUncommitted).toHaveBeenLastCalledWith({ cwd: "foo" });
+    }
     expect.assertions(1);
   });
 });