+++ utils/check-working-tree/lib/check-working-tree.js
@@ -2,10 +2,12 @@
 
 const describeRef = require("@lerna/describe-ref");
 const ValidationError = require("@lerna/validation-error");
+const collectUncommitted = require("@lerna/collect-uncommitted");
 
 module.exports = checkWorkingTree;
+module.exports.mkThrowIfUncommitted = mkThrowIfUncommitted;
 module.exports.throwIfReleased = throwIfReleased;
-module.exports.throwIfUncommitted = throwIfUncommitted;
+module.exports.throwIfUncommitted = mkThrowIfUncommitted();
 
 function checkWorkingTree({ cwd } = {}) {
   let chain = Promise.resolve();
@@ -17,7 +19,7 @@ function checkWorkingTree({ cwd } = {}) {
     // prevent duplicate versioning
     chain.then(throwIfReleased),
     // prevent publish of uncommitted changes
-    chain.then(throwIfUncommitted),
+    chain.then(mkThrowIfUncommitted({ cwd })),
   ];
 
   // passes through result of describeRef() to aid composability
@@ -33,11 +35,15 @@ function throwIfReleased({ refCount }) {
   }
 }
 
-function throwIfUncommitted({ isDirty }) {
-  if (isDirty) {
-    throw new ValidationError(
-      "EUNCOMMIT",
-      "Working tree has uncommitted changes, please commit or remove changes before continuing."
-    );
-  }
+const EUNCOMMIT_MSG =
+  "Working tree has uncommitted changes, please commit or remove the following changes before continuing:\n";
+
+function mkThrowIfUncommitted(options = {}) {
+  return function throwIfUncommitted({ isDirty }) {
+    if (isDirty) {
+      return collectUncommitted(options).then(uncommitted => {
+        throw new ValidationError("EUNCOMMIT", `${EUNCOMMIT_MSG}${uncommitted.join("\n")}`);
+      });
+    }
+  };
 }