+++ utils/check-working-tree/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## Unpublished
+
+### Features
+
+* Display uncommitted changes in error message when working tree is dirty.
+
 ## [3.13.3](https://github.com/lerna/lerna/compare/v3.13.2...v3.13.3) (2019-04-17)
 
 **Note:** Version bump only for package @lerna/check-working-tree